+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [v0.1.9]
 ### Changes
 - [pileup] When estimating the pass-threshold only use base modification probabilities if the read base is aligned to the reference (don't use soft-clipped and inserts). Use `--include-unmapped` to use all base modification probabilities.
+- [adjust-mods] requires `--ignore`, `--convert`, or `--edge-filter`.
 ### Adds
 - [pileup, extract, sample-probs, summary] Allow narrowing of analysis to specific sites with `--include-bed`. 
 - [summary, sample-probs] Add `--only-mapped` flag that will only report on base modification probabilities if they are mapped to the reference.